# swagger.yaml
basePath: /api
definitions:
  database.CardConfirmation:
    properties:
      address:
        type: string
      currency_id:
        type: integer
      id:
        type: integer
      image:
        items:
          type: integer
        type: array
      user_id:
        type: integer
      verified:
        type: boolean
    type: object
...
  server.ApproveCardConfirmationRequest:
    properties:
      confirmation_id:
        type: integer
    type: object
  server.Balance:
    properties:
      address:
        type: string
      balance:
        type: number
      code:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
  server.Balances:
    properties:
      balances:
        items:
          $ref: '#/definitions/server.Balance'
        type: array
    type: object
  server.Busy:
    properties:
      busy:
        type: boolean
    type: object
  server.CancelCardRequest:
    properties:
      confirmation_id:
        type: integer
    type: object
  server.CancelOrderRequest:
    properties:
      order_id:
        type: integer
    type: object
  server.CardConfirmationsResponse:
    properties:
      confirmations:
        items:
          $ref: '#/definitions/database.CardConfirmation'
        type: array
    type: object
  server.CreateBalanceRequest:
    properties:
      address:
        type: string
      balance:
        type: number
      currency_id:
        type: integer
    type: object
...
  server.ExecuteOrderRequest:
    properties:
      order_id:
        type: integer
    type: object
  server.GetCardConfirmationsResponse:
    properties:
      card_confirmations:
        items:
          $ref: '#/definitions/database.CardConfirmation'
        type: array
    type: object
  server.Order:
    properties:
      address:
        type: string
      amountin:
        type: number
      amountout:
        type: number
      approvepic:
        items:
          type: integer
        type: array
      currin:
        type: string
      currout:
        type: string
      email:
        type: string
      id:
        type: integer
      status:
        type: string
    type: object
...
  server.Orders:
    properties:
      orders:
        items:
          $ref: '#/definitions/server.Order'
        type: array
    type: object
  server.RemoveBalanceRequest:
    properties:
      id:
        type: integer
    type: object
...
  server.UpdateBalanceRequest:
    properties:
      address:
        type: string
      balance:
        type: number
      balance_id:
        type: integer
      currency_code:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Swap Support
    url: http://github.com/d1nch8g/swap
  description: Swap exchange api.
  license:
    name: MIT
    url: https://github.com/d1nch8g/swap/src/branch/main/LICENSE
  termsOfService: http://github.com/d1nch8g/swap
  title: Swap exchange api
  version: "1.0"
paths:
  /admin/cancel-order:
    post:
      parameters:
      - description: Cancel order parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CancelOrderRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Cancel user order and mark it as cancelled
  /admin/change-busy:
    post:
      parameters:
      - description: Busy status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.Busy'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Change busy status for admin operator
...
  /admin/create-balance:
    post:
      parameters:
      - description: Create balance parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateBalanceRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new operator balance
...
  /admin/execute-order:
    post:
      parameters:
      - description: Execute order parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.ExecuteOrderRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Execute order and change operator balances, update busy
  /admin/list-balances:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Balances'
      security:
      - ApiKeyAuth: []
      summary: List operator currency balances
  /admin/remove-balance:
    delete:
      parameters:
      - description: Balance id
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.RemoveBalanceRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove operators balance
...
  /operator/approve-card:
    post:
      parameters:
      - description: Approve card request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.ApproveCardConfirmationRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Mark user credit card as approved
  /operator/cancel-card:
    delete:
      parameters:
      - description: Remove confirmation id
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CancelCardRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove user's card request
  /operator/card-confirmations:
    get:
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CardConfirmationsResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user's card confirmations
  /operator/finished-orders:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Orders'
      security:
      - ApiKeyAuth: []
      summary: Get finished orders bound to specific operator
  /operator/get-card-confirmations:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GetCardConfirmationsResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user credit card approval images with parameters
  /operator/get-orders:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Orders'
      security:
      - ApiKeyAuth: []
      summary: Get active orders bound to specific operator
  /operator/update-balance:
    post:
      parameters:
      - description: Update balance parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.UpdateBalanceRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update operator currency balance
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Token authorization for internal operations
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"